profile: Support yappi clock types
yappi supports both "cpu" clock and "wall" clcok. The default "cpu"
clock is handy when you want to see when time is spent running code. The
wall clock is handy when you want to understand when time is spent
running or sleeping.

See test_cpu_clock and test_wall_clock if the difference between the
clock is not clear.

Change-Id: I64069de315a0bad5cb19365b43edf79aceeedca1
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27816
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Saggi Mizrahi <smizrahi@redhat.com>