profile: Allow disabling threads when using the decorator
yappi supports profiling all threads (default) or only the thread where
the profiler was started from. This patch add this option to the profile
decorator. This is is probably not going to be used very much, but
useful when you need it.

New threading tests verify that yappi indeed work correctly with other
threads, even those that started before yappi was started.

Change-Id: I6608a831e0ac2d347d9b2f00458e570da31fdafd
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27894
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-by: Saggi Mizrahi <smizrahi@redhat.com>