BindingXmlRPC - do_PUT to return created task id in header
Currently the execution result of "download" to image is being returned
as json written to the response message body, this patch changes it so that
in case of failure the failure details (error code and message) would be
returned as json in the message body while in case of succesful
task creation the task id would be returned in a header.
this is needed so that when the abillity to upload an image from vdsm
would be added, both operations will return a similar result (as in the
"upload" operation the response message body would contain the contain the
image data so we'll need to return the task id within the header).

Change-Id: I25b13f44a65cb2c794c458b03ba361bf0af92120
Signed-off-by: Liron Aravot <laravot@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27997
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>