ifcfg: Prevent ifup from failing due to existing dhclient
When configuring a device with bootproto=dhcp it is possible that
the device, if pre-existent, had a running dhclient instance that
would make ifup fail. This is especially true in the oVirt
deployment flow when we configure the management network over
a device that very likely already had a dhclient running.

Up until now this has not hit us hard (though it has polluted logs)
due to the fact that we ifup dhcp in asyncronous mode and return
success before the failure.

Bug-Url: https://bugzilla.redhat.com/1100264
Change-Id: Icc4116e1064a00aa891003c9c8fdaea61dabca2c
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/28038
Reviewed-by: Martin Pavlik <xpavli44@gmail.com>
Tested-by: Martin Pavlik <xpavli44@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>