iproute2: stop non-vdsm dhclient before we run our own
It is possible that a device that we re-use, like for example a nic
or a bond in the installation flow could have a dhclient process
running for them. That would effectively make the vdsm dhclient run
fail due to dhclient's detection of it having a 'dhcp agent' already
running. This patch solves it by pre-emptively telling dhclient to
kill clients for existent devices that we are about to attempt to
configure.

Bug-Url: https://bugzilla.redhat.com/1100264
Change-Id: Id9bbca6e2039f5c5ef4ee6172a2c0d9d35ebdbb0
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/28039
Reviewed-by: Dan Kenigsberg <danken@redhat.com>