caps: add reporting of all ipv4 addresses
With this patch, we add a new field to the reported device
information. This new field is 'ipv4addrs', that just like the
current 'ipv6addrs' reports all the addresses of the device (each
for its own address family). With this new reporting, now 'addr'
and 'netmask' are specifically reporting the primary address.

Change-Id: I364a2737182b3a3cac914f1bbd3d707e1346f40f
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/28063
Reviewed-by: Dan Kenigsberg <danken@redhat.com>