hsm: unify vm ovf management lock
Taking one lock per VM definition could be extremely slow (especially
when the engine sends 100 ovf). In one case it took more than 3 minutes
to take 76 locks. This patch unifies the VM locks into a generic one.
There was no real advantage in the current implementation, in fact if
by any chance updateVM was stuck trying to persist a VM definition it
is understood that also removeVM would get stuck at file-system level.
Moreover storing 100 ovf is a quite fast operation (if it doesn't get
stuck) and it is not supposed to starve removeVM.
Multiple removeVM calls are still executed in parallel because the
generic lock is acquired in shared mode.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1103844
Relates-To: https://bugzilla.redhat.com/show_bug.cgi?id=1100527
Change-Id: Ia857d4afa25e7eebd59fb10ef31e98c25caa9928
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/28068
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/28379
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>