Setting enum for isconfigured return value and change override semantics
isconfigured now returns CONFIGURED, NOT_CONFIGURED and NOT_SURE. This to
ease and clarify the override semantics.
If module returns CONFIGURED it means that also on force vdsm won't need to
reconfigure it, as the module configured properly. Although if configured
returns NOT_SURE, it means, vdsm already configured the module, but on
force we desire to override those configuration to defaults.

Change-Id: I58b2eef787a90073a06caf88b6847f34fbd042ed
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1088805
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/28007
Reviewed-by: Dan Kenigsberg <danken@redhat.com>