LiveMerge: Extend internal volumes during live merge
During a live merge operation data is written to the target of the
commit (the base volume).  This means that for block-based images, we
must monitor the base volume and the leaf volume for extension.

Since libvirt is not providing watermark information for internal
volumes yet, return an empty set for now until we get the API we need.

The libvirt bug for the missing API is:
https://bugzilla.redhat.com/show_bug.cgi?id=1041569

Change-Id: I1a5b11e7da185a699028b6127066cd01de010a0d
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/28597
Reviewed-by: Dan Kenigsberg <danken@redhat.com>