core: GET requests - use Range header
This patch replaces the use of the custom Size header with use of http
Range header.
(http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.35)
Note that the use is similar to the use that was done with the Size header.
The motivation is to use the standard http header (instead of supporting a
custom header) and not to fully comply with the spec (as there are existing
gaps).

After this change the Range header is mandatory to issue a get request for an
image and support ranges between zero and specified last byte position.

Change-Id: I8164867347b1cf800efd2a78cc98dbc10c02ee0d
Related-To: https://bugzilla.redhat.com/show_bug.cgi?id=1037500
Signed-off-by: Liron Aravot <laravot@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/28465
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29553
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>