virt: do not access devices directly outside Vm
VmStatsThread used to access directly the _devices and conf
fields of vm.Vm.
This patch does a little cleanup by introducing helper getters
to mediate the access to the devices for the Vm clients,
VmStatsThread being the first.

this patch add little benefit on its own but paves the road
to extract VmStatsThread from vm.py and put it in sampling.py.

Change-Id: I4313da06153fd9f2db90315d16b242352a127d01
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/28478
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>