net models: simplify error condition
An attempt to add a bonding device with no nics ends with a different
error, depending whether the bond is already define in the kernel or
not. This surprising behaviour means that the functional test
testFailWithBadParam fails if BONDING_NAME is removed from the kernel.

This patch unifies the error code, returning ERR_BAD_PARAMS in both
cases.

Change-Id: I557ba2dcecbfa40f9268e8c3da0562f491a32c3c
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/28497
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>