virt: Specify snapshot type
Libvirt 1.2.2 introduced the 'type' attribute to the source element in a
domainsnapshot XML document.  This attribute allows you to specify
'block' as the type (which we should be doing on block domains).
Leaving this out creates a snapshot of type 'file' which has some
troubling side effects.  For one, it changes the disk type from block to
file.  This interferes with libvirt's write watermark detection causing
the wrong value to be returned.  In turn, vdsm will extend the drive
until it consumes the entire block storage domain.  See bug 1091094 [1].

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1091094

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1091094
Change-Id: I940ef0fb702708bf2158739eeed7be43ccce78d6
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/28531
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>