clarify that scsiKey parameter is not passed on
In this patch I add documentation to make plain that the scsiKey parameter is
a legacy from older code. In the code itself, I changed the parameter
name so that this is clear to the next developer that comes along, and
hopefully will save him/her a wild goose chase.

Change-Id: I5edc98058f2332deb6ee040879b4f7c635aefbc2
Signed-off-by: Yoav Kleinberger <ykleinbe@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/28560
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>