sd: Fix wrong usage of properties on old style class
StorageDomain is an old style class but usess properties, which are
supported *only* for new style classes. This is a nasty corner in Python
where you can define @property on old style classes and they seems to
work in some of the cases, and break in others. The current code happen
to work because we never try to assign to the read only properties.

The patch change StorageDomain to be a new style class, and fix the
tests assuming an old style class.

For more info see:
https://docs.python.org/2.6/library/functions.html#property

Change-Id: I89aee24f1fc472bf6425e4194cea58c7f7b52dc8
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/28655
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>