vdscli: avoid ssl-specific localhost heuristic
Unlike it used to, Vdsm no longer checks its client's identity versus
the content of the client certificate.

Therefore, there is no need for the too-clever mechanism of guessing the
local host adderess from the ssl certificate. That mechanism is known to
have caused pains when the subject name of the certificate became
non-resolvable for whatever reason (Cf. http://gerrit.ovirt.org/28107 ).

Change-Id: I421e466621e2a172f9cf8158913654c49cfecca1
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1111075
Reviewed-on: http://gerrit.ovirt.org/28674
Reviewed-by: Alon Bar-Lev <alonbl@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29182
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>