vm: fix guest agent stats field
In commit 068fc28d I mistakenly removed an explicit int() conversion
because I overlooked the code, and thought the result was going to
be an int() anyway.

I was wrong, and the result is a float, which breaks the schema which
dictates the affected field to be an uint.

This patch restores the explicit conversion to make the code compliant
with the schema again.

Change-Id: Iab5ab0475e5fdf560ce07cdc1868690d3eceb125
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/28781
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Martin Sivák <msivak@redhat.com>
Reviewed-by: Kobi Ianko <kobi@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>