domainMonitor: Introduce cancellation points
While monitoring domain, the monitor perform some operations that we
would like to avoid if the monitor thread was stopped while running
the _monitorDomain method. Example operations are accessing shared
storage, which may block for long time, or acquiring a host id.

This patch adds the @cancelpoint decorator, that can be used to mark
methods as cancellation points. Some of the domain monitor thread
methods are marked as cancellation points, ensuring that a monitor will
stop as soon as possible when stopped.

Change-Id: I0a965758736a0b9fa7ac7c2e105bcbbfb04163b8
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/28721
Reviewed-by: Yoav Kleinberger <ykleinbe@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>