vdsm: jsonrpc reboot issue
When rebooting a host using power management UI there was an issue that
proxy host failed to start it. The issue was that jsonrpc code expects
power attribute as it is specified in the schema but it was missing when
running other commands than status. 

Now for commands on and off we provide power attribue as 'unknown' which
fixes the issue.


Change-Id: Ie7be1c00c35f4c2b4c99ed01081020198c7a566d
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: http://gerrit.ovirt.org/28839
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>