netinfo: Report bonding options as a dictionary of strings
The values of bonding options are read from the kernel
as one or two strings. If there is only one string,
the value is actually numeric. If there are two,
the value is an enumeration, with a symbolic name in
the first string, a numeric code in the second string.

To simplify the API from the current state (reporting
the options as "raw" arrays of one or two strings),
the only value reported now is the number in both cases,
left in its textual representation.

The key was changed from 'options' to 'opts'
to mirror 'opts' in _bridgeinfo.

A bond info dictionary is thus changed to (abriged):
'bond4': {
    'cfg': {'BONDING_OPTS': 'miimon=150 mode=3'},
    'opts': {'miimon': '150', 'mode': '3'}
}

Change-Id: I1d29618c6d19d8c7f36fa674043488480dc346fb
Bug-Url: https://bugzilla.redhat.com/987813
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/28845
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>