ifcfg: use oVirt node's to remove files
Doing this, when flushing the old network configuration, oVirt node
will take care to unmount and unpersist the files, as well as
removing them.

The bug was that ifcfg.flush runs at boot and tries to remove all the
ifcfg files, but those that are persisted are untouchable due to
the bindmount.

Bug-Url: https://bugzilla.redhat.com/1104035
Change-Id: Ia9f8ef708c7681bb347e4ee6fba512b2cfa49611
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/28879
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Tested-by: Douglas Schilling Landgraf <dougsland@redhat.com>