ifcfg: perform blocking dhcp when in rollback
When rolling back configuration, be it on boot or due to loss of
connectivity, it is best to know if something goes wrong as early as
possible. One important part of this is knowing if ifup (and
specially its dhclient component) succeeded.

When booting up, parallel dhcp initializations are troublesome and
can sometimes fail, it is better to perform network restoration
serially as with ifcfg persistence (initscripts serializes its device
ifups).

Change-Id: I5b9614ed8fb2404ccfffb72e08be4f76eb366db9
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/28935
Tested-by: Martin Peřina <mperina@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>