Fix exception in _getUserCpuTuneInfo and _getCpuCount
If the cpu tune metadata is not present in a VM's domain xml then
_getUserCpuTuneInfo will fail with the following exception:

Traceback (most recent call last):
  File "/usr/share/vdsm/virt/vm.py", line 2846, in _getRunningVmStats
    decStats = self._vmStats.get()
  File "/usr/share/vdsm/virt/vm.py", line 571, in get
    self._getUserCpuTuneInfo(stats)
  File "/usr/share/vdsm/virt/vm.py", line 421, in _getUserCpuTuneInfo
    if eInfo['vcpuLimit']:
KeyError: 'vcpuLimit'

This aborts the rest of VM stats collection and causes lots of problems
for other flows which depend on VM stats (such as Live Merge).

The same applies to _getCpuCount where the value can be empty in case
libvirt has issuess accessing the cgroup controllers (currently this
happens on Fedora 20).

Fixed by fist checking the availability of the key.

Change-Id: Ifd5fa19f7985adab6959d6011d1eaf467454d925
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/28966
Tested-by: Martin Sivák <msivak@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>