Removed runtime dependencies
On centos there is old jackson library so we have decided to remove
runtime dependency on jackson for both rhel and centos. When running as
part of the engine the library is provided by jboss AS. Standalown usage
will be not supported out of the box.

Jackson dependency is conditional based on the value of
jackson_deps_available. If the dept is not available we use maven to
build the project.

Change-Id: Ic14acae23b64a21395eb591a2dd9fd804e4a2728
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Signed-off-by: Saggi Mizrahi <ficoos@gmail.com>