domainMonitor: Fix unsafe iteration of domain monitor status
When we generate repoStat response, we used to iterate over the
monitored domains, getting the status for each monitor. This naïve
approach does not consider that a monitor may be removed from the domain
monitor dict during the iteration.

Since the only usage now is getting the status for all domains, I
changed getStatus to return an iterator returning all domains status in
a safe manner.

Change-Id: I083c4b1925ee529c0993a291cad6076539d4ea66
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29009
Reviewed-by: Yoav Kleinberger <ykleinbe@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>