domainMonitor: Fix unsafe status handling
DomainMonitor status was updated in a unsafe way, assuming that another
thread will never ask the status in the middle of an update. The code
was written as if creating a new status object is expensive operation
that should be avoided.

Instead of having long living status object in the monitor, I create new
forozen copy each monitor interval (10 seconds). The frozen copy ensure
that careless developers trying to modify a status will get what they
deserve.

Change-Id: If4affb67b6e65382dc4e55ebad9443e2722f2773
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29014
Reviewed-by: Yoav Kleinberger <ykleinbe@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>