virt: migration: simplify the remote port handling
When the destination port for migrations is not explicitely
given, the current code in _setupVdsConnection does a clumsy
gimmick to get it from the BindingXMLRPC internals.

This unnecessarily makes the coupling stricter.
Instead, just use the port specified as per configuration,
which is the same setting the XMLRPC bindings are using
anyway.

Change-Id: Iee19377c6387738d7c4d4a898b6e716a1eeb884c
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29048
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>