sslutils: Document M2Crypto threading initialization
As documented and painfully recognized long ago in, using M2Crypto in a
multi-threaded server such as Vdsm requires explicit initialization.

M2Crypto was initialized but the code was not docuemnted properly. This
lead to the initialization being removed in this patch:
http://gerrit.ovirt.org/#/c/28858/4/lib/vdsm/verifyingtransport.py

Hopefully, improved documentation and moving it to the sslutils module
will prevent such attempts in the future.

Change-Id: I8e5f8dc4662e16ba7e39f60336819750e1366124
Relates-to: http://www.heikkitoivonen.net/m2crypto/api/M2Crypto.threading-module.html
Relates-to: https://bugzilla.redhat.com/482420
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29104
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>