fencing: Introduce getHostLeaseStatus API
When fencing another host, we like to check if the other host has access
to storage, to prevent unwanted fencing of a lively host. This patch
adds an internal API for getting another host status on monitored
domains.

The new getHostLeaseStatus API returns a dictionary of monitored domains
UUIDs and host lease status for each domain, as reported by the cluster
lock.  The caller will apply a fencing policy using this data to decide
if host is lively enough to prevent fencing.

Here is an example result:

    {'04e604d3-71eb-41e0-a8a3-74404bdd9d75': 'live',
     '09049dc6-2007-492e-afd9-3ec201775b2a': 'dead',
     '20501ea5-cf6b-453a-b23d-a708c28cec1f': 'live',
     '2b02c759-2e4a-4cf0-9b8e-37a6ec77f821': 'unavailable',
     ...
     'fc348be8-18f4-42ff-8421-9043203e13a6': 'live'}

See clusterlock.py for the possible statuses and their meaning.

Change-Id: Iccd62e58a194aa0ceb0f5e2503b8ec7e4349971b
Bug-Url: https://bugzilla.redhat.com/1110172
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/28873
Reviewed-by: Martin Peřina <mperina@redhat.com>
Tested-by: Martin Peřina <mperina@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/31148
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>