detach: Support force detach on Storage Domains with old pools.
Force detach of storage domain only being supported on storage domains
which are part of pool.  For supporting import Data Storage Domain
feature, we should also support force detach of Storage Domain which are
not connected to any pool.

The force detach will clear the SP_UUID meta data of the imported
Stroage Dimain, when a BLANK_UUID will be passed from the client
and will enable the engine to attach it to other DC in ifferent setups.

Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>
Change-Id: I4a6fb01775f7c834a3466e45b9b7ed9b4bd5c3be
Reviewed-on: http://gerrit.ovirt.org/29042
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>