vdsm: lastclient info in jsonrpc
In xml binding there is a hack for getCapabilites which before
sending response adds listclient and lastClientIface.

We can not modify API.py to provide this information in a common place
because it is meaningless for mom thread , or any other in-process API
consumer.

We provide similar hack for jsonrpc at the level where we have enough
information to send back required information.

Bug-Url: https://bugzilla.redhat.com/1099500
Change-Id: I402e15cb05f89a98dab14491d9da5985335e095e
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: http://gerrit.ovirt.org/28817
Reviewed-by: Saggi Mizrahi <smizrahi@redhat.com>