tc: replace mirred filter parsing with a generic filter parser
For now, we only need to parse filters with mirred actions. This
must change in order to implement the Quality of Service feature,
which will require parsing classes, queuing disciplines and other
kinds of filters.

This patch adds a generic parsing filter and lays the structure
for iterating over other Traffic Control entities.

Change-Id: I95aa9aaaeb0f8e612734b32a578e46fade7fdd9c
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29211
Reviewed-by: Dan Kenigsberg <danken@redhat.com>