vm: hyperv: make hw clock friendlier to windows
This patch sets the local time as windows reportedly
expect. According to libvirt docs:

" Most operating systems expect the hardware clock to
be kept in UTC, and this is the default.
Windows, however, expects it to be in so called 'localtime'."
http://libvirt.org/formatdomain.html#elementsTime

This is the reccommended setting for windows and reportedly
has stability benefits for live migrations.

The new setting is enabled on top of the hyperv optimizations,
and is disabled by default for backward compatibility.

Change-Id: I84a8cfded683a42826b6dabb67b4a5eb456594d7
Bug-Url: https://bugzilla.redhat.com/1107835
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29233
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>