ifcfg: make default route network be started by sysV
With the change to unified persistence, the networks defined by vdsm
are only started and restored on boot by vdsmd itself. The issue with
this is that if vdsmd fails to start for whichever reason, that may
impede remote access to the machine.

This patch is a transitional improvement (future releases will have
separate vdsm-network and vdsmd services and initializations).

This patch solves the issue by making all the interfaces that are
needed for the network that is defined to carry the default route
(typically the management network) have ONBOOT=yes. Setting this
makes the initscripts network service configure the network itself.
So for the default route network, the boot process is going to:
    a) sysV or systemd start network.service,
    b) network.service goes over all the ifcfg files and ifups those
       with ONBOOT=yes
    c) vdsmd start will call vdsm-network-restore-config which will:
        + flush ifcfg files
        + call setupNetworks to restore all the defined networks.

It's easy to see that if (c) does not happen we will be left with
only the default route network, so remote access will still be
possible.

Change-Id: Iaf656958e8ddfdd7e8be9a8bcb51962e073ae26a
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29704
Reviewed-by: Dan Kenigsberg <danken@redhat.com>